tidy(ui): organise canvas object classes
