+++ invokeai/frontend/web/src/features/controlLayers/konva/CanvasObject/CanvasObjectImage.ts
