+++ invokeai/frontend/web/src/features/controlLayers/konva/CanvasObject/CanvasObjectRect.ts
